add amazon s3

+++ utils/image.go
@@ -267,10 +267,22 @@ func (i *ImageType) saveFile() (err error) {
 		return errors.New("problem saving file")
 	}
 
-	err = UploadGCS(imagefile, fmt.Sprintf("src/%s", i.Filename))
-	if err != nil {
-		os.RemoveAll(imagefile)
-		return
+	// upload the file to google if capability is set
+	if Services.Storage.Google {
+		err = UploadGCS(imagefile, fmt.Sprintf("src/%s", i.Filename))
+		if err != nil {
+			os.RemoveAll(imagefile)
+			return
+		}
+	}
+
+	// upload the file to amazon if capability is set
+	if Services.Storage.Amazon {
+		err = UploadS3(imagefile, fmt.Sprintf("src/%s", i.Filename))
+		if err != nil {
+			os.RemoveAll(imagefile)
+			return
+		}
 	}
 
 	return
@@ -324,11 +336,24 @@ func (i *ImageType) createThumbnail() (err error) {
 	i.ThumbWidth = img.Width
 	i.ThumbHeight = img.Height
 
-	err = UploadGCS(thumbfile, fmt.Sprintf("thumb/%s", i.Thumbnail))
-	if err != nil {
-		os.RemoveAll(thumbfile)
-		os.RemoveAll(imagefile)
-		return
+	// upload the file to google if capability is set
+	if Services.Storage.Google {
+		err = UploadGCS(thumbfile, fmt.Sprintf("thumb/%s", i.Thumbnail))
+		if err != nil {
+			os.RemoveAll(thumbfile)
+			os.RemoveAll(imagefile)
+			return
+		}
+	}
+
+	// upload the file to amazon if capability is set
+	if Services.Storage.Amazon {
+		err = UploadS3(thumbfile, fmt.Sprintf("thumb/%s", i.Thumbnail))
+		if err != nil {
+			os.RemoveAll(thumbfile)
+			os.RemoveAll(imagefile)
+			return
+		}
 	}
 
 	return
